+++ String Algorithms/Finding Patterns.cpp
@@ -59,7 +59,7 @@ bool query(){
 
 int main(){
     scanf(" %s %d", S, &K);
-    N = strlen(S);
+    N = (int) strlen(S);
     init();
 
     for(int i = 0; i < N; i++)
@@ -67,7 +67,7 @@ int main(){
 
     for(int i =  0; i < K; i++){
         scanf(" %s", T);
-        M = strlen(T);
+        M = (int) strlen(T);
         printf("%s\n", query() ? "YES" : "NO");
     }
 }
