+++ String Algorithms/Repeating Substring.cpp
@@ -74,7 +74,7 @@ void dfs(int u = 0){
 
 int main(){
     scanf(" %s", S);
-    N = strlen(S);
+    N = (int) strlen(S);
 
     init();
     for(int i = 0; i < N; i++)
